backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
aiosqlite==0.18.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.11.1
pytest==7.4.0
httpx==0.24.1
pytest-asyncio==0.21.0
email-validator==1.3.1
argon2-cffi==21.3.0
python-pptx==0.6.21
beautifulsoup4==4.12.2
